make shared struct int fields int64 so that they don't need to be converted for other protocols

+++ shared/reddit_account.go
@@ -8,14 +8,14 @@ type ShRedditAccount struct {
 }
 
 type AccountData struct {
-	CommentKarma int    `json:"comment_karma"`
+	CommentKarma int64  `json:"comment_karma"`
 	HasMail      bool   `json:"has_mail"`
 	HasModMail   bool   `json:"has_mod_mail"`
 	ID           string `json:"id"`
-	InboxCount   int    `json:"inbox_count"`
+	InboxCount   int64  `json:"inbox_count"`
 	IsFriend     bool   `json:"is_friend"`
 	IsGold       bool   `json:"is_gold"`
-	LinkKarma    int    `json:"link_karma"`
+	LinkKarma    int64  `json:"link_karma"`
 	ModHash      string `json:"mod_hash"`
 	Name         string `json:"name"`
 	Over18       bool   `json:"over_18"`
